Move vanish-related methods to VanishCommand.java
