+++ src/main/java/de/shiewk/smoderation/command/VanishCommand.java
@@ -2,6 +2,8 @@ package de.shiewk.smoderation.command;
 
 import de.shiewk.smoderation.SModeration;
 import de.shiewk.smoderation.util.PlayerUtil;
+import it.unimi.dsi.fastutil.objects.ObjectArrayList;
+import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabExecutor;
@@ -10,6 +12,9 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
+
+import static de.shiewk.smoderation.SModeration.*;
+import static net.kyori.adventure.text.Component.text;
 
 public class VanishCommand implements TabExecutor {
 
@@ -22,7 +27,7 @@ public class VanishCommand implements TabExecutor {
             player = (Player) sender;
         }
         if (player != null){
-            SModeration.toggleVanish(player);
+            toggleVanish(player);
             return true;
         } else {
             return false;
@@ -35,5 +40,53 @@ public class VanishCommand implements TabExecutor {
             return PlayerUtil.listPlayerNames();
         }
         return List.of();
+    }
+
+    private static final ObjectArrayList<Player> vanishedPlayers = new ObjectArrayList<>();
+
+    public static void toggleVanish(Player player){
+        final boolean newStatus = !isVanished(player);
+        if (newStatus){
+            vanishedPlayers.add(player);
+            for (CommandSender sender : SModeration.container.collectBroadcastTargets()) {
+                sender.sendMessage(CHAT_PREFIX.append(
+                        player.displayName()
+                                .colorIfAbsent(SECONDARY_COLOR)
+                ).append(
+                        text()
+                                .content(" vanished.")
+                                .color(PRIMARY_COLOR)
+                ));
+            }
+            player.sendMessage(CHAT_PREFIX.append(text("You are now vanished.").color(PRIMARY_COLOR)));
+            player.setVisibleByDefault(false);
+            for (Player onlinePlayer : Bukkit.getOnlinePlayers()) {
+                if (onlinePlayer.hasPermission("smod.vanish.see")){
+                    onlinePlayer.showEntity(PLUGIN, player);
+                }
+            }
+        } else {
+            vanishedPlayers.remove(player);
+            for (CommandSender sender : container.collectBroadcastTargets()) {
+                sender.sendMessage(CHAT_PREFIX.append(
+                        player.displayName()
+                                .colorIfAbsent(SECONDARY_COLOR)
+                ).append(
+                        text()
+                                .content(" re-appeared.")
+                                .color(PRIMARY_COLOR)
+                ));
+            }
+            player.sendMessage(CHAT_PREFIX.append(text("You are no longer vanished.").color(PRIMARY_COLOR)));
+            player.setVisibleByDefault(true);
+        }
+    }
+
+    public static boolean isVanished(Player player){
+        return vanishedPlayers.contains(player);
+    }
+
+    public static ObjectArrayList<Player> getVanishedPlayers() {
+        return vanishedPlayers.clone();
     }
 }
